requirements.txt
           
beautifulsoup4==4.12.2
httpx==0.23.0
matplotlib==3.8.0
numpy==1.26.3
pandas==2.1.4
python-dotenv==0.21.0